requirements.txt
faker==13.6.0
pyspark==3.2.1
hypothesis_jsonschema==0.22.0
hypothesis==6.46.2
pandas==1.4.2
pytest==7.1.2
requests==2.27.1
tenacity==8.0.1
responses==0.21.0
psycopg2==2.9.3
google-cloud-storage==2.3.0 
boto3==1.23.10
moto==3.1.11
apache-airflow==2.3.2